requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
sqlite3==2.6.0